+++ sdk/appconfiguration/azure-data-appconfiguration/src/samples/java/com/azure/data/appconfiguration/ConfigurationAsyncClientJavaDocCodeSnippets.java
@@ -6,12 +6,12 @@ package com.azure.data.appconfiguration;
 import com.azure.core.http.MatchConditions;
 import com.azure.core.util.Configuration;
 import com.azure.data.appconfiguration.models.ConfigurationSetting;
+import com.azure.data.appconfiguration.models.ConfigurationSettingsFilter;
 import com.azure.data.appconfiguration.models.ConfigurationSnapshot;
 import com.azure.data.appconfiguration.models.SettingFields;
 import com.azure.data.appconfiguration.models.SettingSelector;
 import com.azure.data.appconfiguration.models.SnapshotFields;
 import com.azure.data.appconfiguration.models.SnapshotSelector;
-import com.azure.data.appconfiguration.models.SnapshotSettingFilter;
 import reactor.util.context.Context;
 
 import java.time.Duration;
@@ -344,9 +344,9 @@ public class ConfigurationAsyncClientJavaDocCodeSnippets {
     public void beginCreateSnapshotMaxOverload() {
         ConfigurationAsyncClient client = getAsyncClient();
         // BEGIN: com.azure.data.appconfiguration.configurationasyncclient.beginCreateSnapshotMaxOverload
-        List<SnapshotSettingFilter> filters = new ArrayList<>();
+        List<ConfigurationSettingsFilter> filters = new ArrayList<>();
         // Key Name also supports RegExp but only support prefix end with "*", such as "k*" and is case-sensitive.
-        filters.add(new SnapshotSettingFilter("{keyName}"));
+        filters.add(new ConfigurationSettingsFilter("{keyName}"));
         String snapshotName = "{snapshotName}";
         client.beginCreateSnapshot(snapshotName, new ConfigurationSnapshot(filters)
                 .setRetentionPeriod(Duration.ofHours(1)))
@@ -393,8 +393,8 @@ public class ConfigurationAsyncClientJavaDocCodeSnippets {
                     // empty value other than the `fields` specified in the request.
                     System.out.printf("Snapshot name=%s is created at %s, snapshot status is %s.%n",
                         getSnapshot.getName(), getSnapshot.getCreatedAt(), getSnapshot.getStatus());
-                    List<SnapshotSettingFilter> filters = getSnapshot.getFilters();
-                    for (SnapshotSettingFilter filter : filters) {
+                    List<ConfigurationSettingsFilter> filters = getSnapshot.getFilters();
+                    for (ConfigurationSettingsFilter filter : filters) {
                         System.out.printf("Snapshot filter key=%s, label=%s.%n", filter.getKey(), filter.getLabel());
                     }
                 });