+++ sdk/appconfiguration/azure-data-appconfiguration/src/samples/java/com/azure/data/appconfiguration/ListSnapshots.java
@@ -8,9 +8,9 @@ import com.azure.core.util.Configuration;
 import com.azure.core.util.Context;
 import com.azure.core.util.polling.SyncPoller;
 import com.azure.data.appconfiguration.models.ConfigurationSetting;
+import com.azure.data.appconfiguration.models.ConfigurationSettingsFilter;
 import com.azure.data.appconfiguration.models.ConfigurationSnapshot;
 import com.azure.data.appconfiguration.models.SnapshotSelector;
-import com.azure.data.appconfiguration.models.SnapshotSettingFilter;
 
 import java.time.Duration;
 import java.util.ArrayList;
@@ -44,9 +44,9 @@ public class ListSnapshots {
         ConfigurationSetting setting2 = client.setConfigurationSetting("TestKey2", null, "value2");
         System.out.printf("[SetConfigurationSetting] Key: %s, Value: %s.%n", setting2.getKey(), setting2.getValue());
         // 1. Prepare the snapshot filters
-        List<SnapshotSettingFilter> filters = new ArrayList<>();
+        List<ConfigurationSettingsFilter> filters = new ArrayList<>();
         // Key Name also supports RegExp but only support prefix end with "*", such as "k*" and is case-sensitive.
-        filters.add(new SnapshotSettingFilter("Test*"));
+        filters.add(new ConfigurationSettingsFilter("Test*"));
 
         // 1. Create first snapshot
         String snapshotNameTest = "{snapshotNameInTest}";
@@ -65,9 +65,9 @@ public class ListSnapshots {
         ConfigurationSetting setting4 = client.setConfigurationSetting("ProductKey2", null, "value2");
         System.out.printf("[SetConfigurationSetting] Key: %s, Value: %s.%n", setting2.getKey(), setting2.getValue());
         // 2. Prepare the snapshot filters
-        List<SnapshotSettingFilter> filters2 = new ArrayList<>();
+        List<ConfigurationSettingsFilter> filters2 = new ArrayList<>();
         // Key Name also supports RegExp but only support prefix end with "*", such as "k*" and is case-sensitive.
-        filters.add(new SnapshotSettingFilter("Product*"));
+        filters.add(new ConfigurationSettingsFilter("Product*"));
 
         // 2. Create second snapshot
         String snapshotNameProduct = "{snapshotNameInProduct}";