+++ sdk/appconfiguration/azure-data-appconfiguration/CHANGELOG.md
@@ -15,6 +15,7 @@ Note: Below breaking changes only affect the version `1.5.0-beta.1`.
 - Renamed `ConfigurationSettingSnapshot`'s properties name,
   - `size` to `sizeInBytes`,
   - `compositionType` to `snapshotComposition`.
+- Renamed `SnapshotSettingFilter` to `ConfigurationSettingsFilter`.
 - Added `azure-core-experimental` as a dependency and replaced usage of `CreateSnapshotOperationDetail` by `PollResult` 
   and removed `CreateSnapshotOperationDetail` class.
 - Replaced